refactored to use double instead of int for upload size

+++ src/main/java/hr/algebra/photoapp_designpatterns_galic/controller/rest/UploadTestController.java
@@ -16,7 +16,7 @@ public class UploadTestController {
 
     //http://localhost:8081/test/upload?size=6
     @RequestMapping(value = "/test/upload", method = {RequestMethod.GET, RequestMethod.POST})
-    public ResponseEntity<String> testUpload(@RequestParam("size") int uploadSizeMb) {
+    public ResponseEntity<String> testUpload(@RequestParam("size") double uploadSizeMb) {
         try {
             consumptionService.recordUpload(uploadSizeMb);
             return ResponseEntity.ok("Upload recorded: " + uploadSizeMb + " MB");
